Move workflow API to agents/workflows subpath
Move workflow-related exports into a dedicated agents/workflows subpath. Renamed packages/agents/src/workflow.ts to workflows.ts, added the ./workflows export entry to package.json, removed workflow re-exports from index.ts, and updated imports across docs, examples, and tests to use "agents/workflows". Added a changeset for the patch release.
