+++ packages/agents/src/index.ts
@@ -3220,32 +3220,3 @@ export class StreamingResponse {
     this._connection.send(JSON.stringify(response));
   }
 }
-
-// ==========================================
-// Workflow Integration Exports
-// ==========================================
-
-export { AgentWorkflow, WorkflowRejectedError } from "./workflow";
-export type {
-  AgentWorkflowEvent,
-  AgentWorkflowStep,
-  DefaultProgress,
-  WorkflowCallback,
-  WorkflowCallbackType,
-  WorkflowProgressCallback,
-  WorkflowCompleteCallback,
-  WorkflowErrorCallback,
-  WorkflowEventCallback,
-  WaitForApprovalOptions,
-  ApprovalEventPayload
-} from "./workflow";
-
-export type {
-  WorkflowStatus,
-  WorkflowTrackingRow,
-  RunWorkflowOptions,
-  WorkflowEventPayload,
-  WorkflowInfo,
-  WorkflowQueryCriteria,
-  WorkflowPage
-} from "./workflow-types";
